server/requirements.txt
asgiref==3.8.1
cffi==1.17.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
greenlet==0.4.13
hpy==0.9.0
readline==6.2.4.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
